requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-dotenv==1.4.2
Pillow==9.5.0
celery==5.3.0
redis==4.5.5
factory-boy==3.2.1
flake8==6.0.0